permission-sets/requirements.txt
aws-cdk.core
aws-cdk.aws_sso
